chore: cleanup empty string type conversion
### what

add helper for handling type conversion from possibly empty string to StringNull

### why

reduce duplication

### testing

unit tests pass
